requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cloudinary==1.29.0
colorama==0.4.4
dj-database-url==0.5.0
Django==4.0.5
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
joblib==1.1.0
newsapi-python==0.2.6
nltk==3.7
numpy==1.22.4
pandas==1.4.2
psycopg2==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
regex==2022.6.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
textblob==0.17.1
tqdm==4.64.0
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
